+++ palmod/Game/Game_REDEARTH_A_DIR.cpp
@@ -2,8 +2,8 @@
 #include "Game_RedEarth_A_DIR.h"
 #include "..\palmod.h"
 
-UINT16 CGame_RedEarth_A_DIR::uRuleCtr = 0;
-int CGame_RedEarth_A_DIR::m_nSavedMode = 31;
+size_t CGame_RedEarth_A_DIR::uRuleCtr = 0;
+size_t CGame_RedEarth_A_DIR::m_nSavedMode = 31;
 
 constexpr auto RedEarth_Arcade_ROM_Base = L"redearth-simm";
 constexpr auto RedEarth_Arcade_ROMSet_3 = 3;
@@ -49,7 +49,7 @@ CGame_RedEarth_A_DIR::CGame_RedEarth_A_DIR(UINT32 nConfirmedROMSize, int nRedEar
     PrepChangeTrackingArray();
 }
 
-sFileRule CGame_RedEarth_A_DIR::GetRuleInternal(int nGameMode, UINT16 nUnitId)
+sFileRule CGame_RedEarth_A_DIR::GetRuleInternal(int nGameMode, size_t nUnitId)
 {
     sFileRule NewFileRule;
     m_nSavedMode = nGameMode;
