+++ palmod/Game/Game_SFIII2_A_DIR.h
@@ -9,18 +9,18 @@ public:
     ~CGame_SFIII2_A_DIR(void) { FlushChangeTrackingArray(); };
 
     //Static functions
-    static UINT16 uRuleCtr;
+    static size_t uRuleCtr;
 
-    static UINT16 GetRuleCtr() { return uRuleCtr; };
+    static size_t GetRuleCtr() { return uRuleCtr; };
     static void ResetRuleCtr() { uRuleCtr = 0; };
 
     static sFileRule GetNextRule();
-    static sFileRule GetRule(UINT16 nRuleId);
+    static sFileRule GetRule(size_t nRuleId);
 
     sFileRule GetNextRuleForSIMMGame() override { return GetNextRule(); };
 
-    BOOL LoadFile(CFile* LoadedFile, UINT16 nSIMMNumber) override { return LoadFileForSIMMGame(LoadedFile, nSIMMNumber); };
-    BOOL SaveFile(CFile* SaveFile, UINT16 nSIMMNumber) override { return SaveFileForSIMMGame(SaveFile, nSIMMNumber); };
+    BOOL LoadFile(CFile* LoadedFile, size_t nSIMMNumber) override { return LoadFileForSIMMGame(LoadedFile, nSIMMNumber); };
+    BOOL SaveFile(CFile* SaveFile, size_t nSIMMNumber) override { return SaveFileForSIMMGame(SaveFile, nSIMMNumber); };
 
     LPCWSTR GetGameName() override { return L"SFIII:2nd Impact (Arcade Rerip)"; };
 };
